feature/nudge-fm5-reminder-notification-solicitors: Refactored code so nudge emails and post is its own service
