+++ build.gradle
@@ -592,7 +592,7 @@ def sonarExclusions = [
   "**uk/gov/hmcts/reform/prl/services.ManageOrderEmailService.java",
   '**uk/gov/hmcts/reform/prl/services/RoleAssignmentServiceForSystemUser.java',
   '**uk/gov/hmcts/reform/prl/services/RoleAssignmentService.java',
-  '**uk/gov/hmcts/reform/prl/services/ServiceOfApplicationService.java'
+  '**uk/gov/hmcts/reform/prl/services/Fm5NotificationService.java'
 ]
 
 sonarqube {
